+++ middleware/multer.js
@@ -1,3 +1,4 @@
 const multer = require('multer')
 const upload = multer({ dest: 'temp/' })
+
 module.exports = upload
